Implemented antialiasing

+++ raytracing/image.h
@@ -5,14 +5,14 @@ class image
 {
 public:
 	image();
-	image(unsigned int width, unsigned int height);
+	image(unsigned long int width, unsigned long int height);
 	image(const image& other);
 	image(image&& other) noexcept;
 	~image();
 	void save_to_file(const std::string& filename);
 	void save_to_ppm(const std::string& filename);
 public:
-	unsigned int width;
-	unsigned int height;
+	unsigned long int width;
+	unsigned long int height;
 	rgb_color* pixels;
 };
